+++ metro-extensions/contribute/impl-compiler-plugin/src/main/kotlin/software/ralf/app/platform/metro/compiler/AppPlatformMetroExtensionsPluginComponentRegistrar.kt
@@ -5,9 +5,6 @@ import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
 import org.jetbrains.kotlin.config.CompilerConfiguration
 import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrarAdapter
-import software.ralf.app.platform.metro.compiler.renderer.ContributesRendererIrExtension
-import software.ralf.app.platform.metro.compiler.robot.ContributesRobotIrExtension
-import software.ralf.app.platform.metro.compiler.scoped.ContributesScopedIrExtension
 
 @AutoService(CompilerPluginRegistrar::class)
 public class AppPlatformMetroExtensionsPluginComponentRegistrar : CompilerPluginRegistrar() {
@@ -16,8 +13,6 @@ public class AppPlatformMetroExtensionsPluginComponentRegistrar : CompilerPlugin
 
   override fun ExtensionStorage.registerExtensions(configuration: CompilerConfiguration) {
     FirExtensionRegistrarAdapter.registerExtension(AppPlatformMetroExtensionsPluginRegistrar())
-    IrGenerationExtension.registerExtension(ContributesRendererIrExtension())
-    IrGenerationExtension.registerExtension(ContributesRobotIrExtension())
-    IrGenerationExtension.registerExtension(ContributesScopedIrExtension())
+    IrGenerationExtension.registerExtension(AppPlatformIrDeclarationGenerationExtension())
   }
 }